// src/app/sistema/page-questoes/enums/tema.ts
export enum Tema {
  AAO_ADAPTED = 'AAO_ADAPTED',
  CATARATA = 'CATARATA',
  CIRURGIA_REFRATIVA = 'CIRURGIA_REFRATIVA',
  CORNEA_CONJUNTIVA_E_ESCLERA = 'CORNEA_CONJUNTIVA_E_ESCLERA',
  CRISTALINO = 'CRISTALINO',
  EMBRIOLOGIA = 'EMBRIOLOGIA',
  ESTRABISMO = 'ESTRABISMO',
  FARMACOLOGIA = 'FARMACOLOGIA',
  GENETICA_OCULAR = 'GENETICA_OCULAR',
  GLAUCOMA = 'GLAUCOMA',
  ICO_PART_1 = 'ICO_PART_1',
  LENTES_DE_CONTATO = 'LENTES_DE_CONTATO',
  NAO_CLASSIFICADA = 'NAO_CLASSIFICADA',
  NEUROFTALMOLOGIA = 'NEUROFTALMOLOGIA',
  OPTICA = 'OPTICA',
  OUTROS = 'OUTROS',
  ORBITA = 'ORBITA',
  PATOLOGIA_OCULAR = 'PATOLOGIA_OCULAR',
  PLASTICA_OCULAR_E_SISTEMA_LACRIMAL = 'PLASTICA_OCULAR_E_SISTEMA_LACRIMAL',
  REFRACAO = 'REFRACAO',
  RETINA = 'RETINA',
  TUMORES_INTRAOCULARES = 'TUMORES_INTRAOCULARES',
  UVEITE = 'UVEITE',
  VISAO_SUBNORMAL = 'VISAO_SUBNORMAL',
  SBRV = 'SBRV',
}